% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remode.R
\name{barplot.remode_result}
\alias{barplot.remode_result}
\title{Barplot for remode_result}
\usage{
\method{barplot}{remode_result}(
  remode_result,
  main = paste("Modes =", remode_result$nr_of_modes),
  density = replace(rep(20, length(remode_result$xt)), remode_result$modes, 50),
  ...
)
}
\arguments{
\item{remode_result}{A list of class `remode_result` containing the output of the `remode` function.}

\item{main}{A character string specifying the main title of the plot. Default is "Modes = (number of modes)".}

\item{density}{A numeric vector specifying the density of shading lines, in lines per inch, for the bars.
Default is 20 for non-mode bars and 50 for mode bars.}

\item{...}{Additional arguments passed to the `barplot` function.}
}
\value{
None. This function is called for its side effects.
}
\description{
The `barplot.remode_result` function provides a way to visualize the output of the `remode`
function using a bar plot, highlighting the identified modes by adjusting the bar density.
}
\examples{
data <- c(80, 90, 110, 70, 90)
result <- remode(data, alpha_correction = "max_modes")
barplot(result)

}
